add create order in matchine

+++ dtos/order_book_dto.go
@@ -7,13 +7,10 @@ import (
 
 // OrderBookDTO ...
 type OrderBookDTO struct {
-	// ID        uint      `json:"id"`
-	// CreatedAt time.Time `json:"created_at"`
-	// UpdatedAt time.Time `json:"updated_at"`
 	models.BaseModel
-	UserID    uint            `json:"user_id"`
+	UserID    uint64          `json:"user_id"`
 	Symbol    string          `json:"symbol"`
-	FundID    uint            `json:"fund_id"`
+	FundID    uint64          `json:"fund_id"`
 	Status    uint            `json:"status"`
 	OrderType string          `json:"order_type"`
 	Side      string          `json:"side"`
